+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## v0.40.2 (2025-04-30)
+
+### Fix
+
+- **ci**: align mcp instrumentation version (#2876)
+
+## v0.40.1 (2025-04-30)
+
+### Fix
+
+- **ci-cd**: add mcp to commitizen (#2875)
+
+## v0.40.0 (2025-04-30)
+
+### Feat
+
+- **instrumentation**: Adding MCP opentelemetry-instrumentation into traceloop (#2829)
+
+## v0.39.4 (2025-04-28)
+
+### Fix
+
+- **sdk**: improve type safety in decorators (#2867)
+
+## v0.39.3 (2025-04-24)
+
+### Fix
+
+- **langchain**: support cached tokens attributes logging (#2830)
+
 ## v0.39.2 (2025-04-18)
 
 ### Fix